bootstrap.sh: Fix SC2163 - null var check already done so silence it
This parameter expansion, `${foo?}`, will return error if `${_var}` is null or
unset... but we already check that before exporting the variable name contained
in `${_var}`. Shellcheck is trying to be helpful, but we actually want to export
the variable contained in `$_var`, not just `$_var` itself.

Shellcheck warning was:

    In bootstrap-scripts/bootstrap.sh line 266:
        [ -n "$(eval echo -n \$$_var)" ] && export ${_var}
                                                   ^-----^ SC2163 (warning): This does not export '_var'. Remove $/${} for that, or use ${var?} to quiet.